requirements.txt
django
djangorestframework
psycopg
psycopg2